requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.33.0
mysql-connector-python==9.0.0
pydantic==2.10.6
